US3T3: Resolve merge conflicts, refactor search into SearchView, update HomeView tile
